backport: don't request reviews from bots
When requesting reviewers, we attempt to request reviews from
those who reviewed the original PR. This can fail because the
CodeQL scanner shows up as a reviewer.

Filter out bots by looking for '[bot]' in the reviewer name.
This catches CodeQL and dependatbot, the two bots we have running
at the time of this commit.