Delete the remote branch when backport workflow completes
This extends our support for cleaning up remote branches.

The original behavior is that we would only delete the remote
branch when backport PRs are merged to one of the release branches.

This means that remote branches aren't cleaned up when:
- PRs get merged to master
- Manual backports get merged to a release branch

These changes ensure that the remote branch is deleted for any
internal contributor's PR, whether it is merged to master or a
release branch, no matter what the name of the remote branch is.