Fix checks for internal authors (#66)
The previous version of the code was checking whether the
author was an external contributor twice, but the 2nd check
happened at a level where we don't have enough information.

The 2nd check is not necessary, since it only occurs in the
code path that is taken for internal authors. This commit
includes a rename to make it clear that these methods are
used for checking PRs from employees.