Covered paths for preferred reviewers
This change will not evaluate a path if that path has already been
covered by a previously picked reviewer.

The current preferred reviewers implementation adds a reviewer for each
unique path defined in the reviews.json file if that path is included in
the change set. This can result in either the same reviewers being added
multiple times, which Github dedups for us, or adds far too many
reviewers when there are a number of paths in the change set.