Automatically add backport labels to docs PRs
Closes gravitational/teleport#15103

We typically need to backport docs PRs to at least one `branch/v[0-9]+`
branch unless a docs PR only applies to the next (not current) major
Teleport release.

Currently, docs authors need to remember to add backport labels to their
PRs so the automatic backport workflow can create branches. If an author
forgets to add the correct labels, someone needs to remember to
complete the backport some other way. If an author forgets to backport a
PR, we need to wait until the next major release before it becomes
visible in the default version of the docs site.

This change adjusts the `*Bot.Label` functionality of the workflow bot
to label docs PRs with `backport/branch/v[0-9]+`  for the current and
past two versions. This way, it's less likely for an author to forget to
label their PR with backport labels.

If a branch does not need to be backported, an author can remove the
labels. If an author forgets to remove the labels, they can always
decline to create a backport branch when we merge the PR, we can delete
the backport branch, etc.

- Add a `GitHub.GetLatestReleaseMajorVersion` function that gets the
  latest Teleport release major version, e.g., `12`
- Add a parameter to `*Bot.labels` for the major version. If this is
  present, and the PR is a docs PR, the workflow will add `backport`
  labels for all supported versions.