Don't count JSON files when checking for large PRs
We sometimes commit large JSON files. These are not code and typically
do not increase the complexity of the PR or the review burden, so don't
count them when checking whether the PR is too large.

This resolves a common issue that requires admin approval to add new
Grafana dashboards to the cloud repo.