Add binary bloat checking (#150)
A new bloat target is added to the bot to determine if binaries
before and after a change cause an increase in size. All changes
less than +1MB are ignored, a warning is emitted for any
changes >=+1MB, and >=+3MB results in an error.

An increase of >=3MB is allowed if an admin leaves a comment on
the PR with the prefix /excludebloat similar to how flaky tests
can be skipped.