Convert bloat check to use previously computed base sizes (#152)
The bot now splits calculating the base artifact sizes and checking
for bloat into two separate workflos. The sizes from the
"binary-sizes" workflow are fed into the "bloat" work flow to use
as the base for comparision. This allows actions to cache a single
file with the artifact sizes instead of a directory with all the
build artifacts both reducing cost and execution time required
to run the bloat check action.