Skip Dependency Review action in merge queue
Various attempts to skip the action were tried in `teleport.e`, but
they all resulted in the Dependency Review check never reporting a
status. This blocked all PRs from landing until the merge queue was
disabled. By defining the skip here instead of in an upstream repo
it should prevent the aforementioned issue.