Allow preferred docs reviewers (#179)
* Allow preferred docs reviewers

The code reviewer configuration allows us to name preferred reviewers
for certain file paths. For docs PRs, the workflow bot does not consult
the preferred reviewer configuration. This means that, even if we add
the `preferredReviewerFor` key to a `codeReviewers` entry with a docs
path as a value, the workflow bot continues to add only `docsReviewers`
to docs PRs.

In this change, the workflow bot gets docs reviewers by:

1. Retrieving the docs reviewer pool from the config (like it does now).
1. Retrieving all preferred reviewers for the file paths included within
   a PR. Since the pool of dedicated docs reviewers is small, the
   workflow bot gets all preferred reviewers for a file path, not just
   one.
1. Merging the two lists of reviewers.

The alternative would be to allow preferred file paths in the docs
reviewer pool, but it makes sense to have a small, consistent set of
docs reviewers who have experience with docs reviews and a larger pool
of subject matter experts for specific file paths from the code reviewer
pool.

Implementation notes:

- Adds an `*Assignments.getAllPreferredReviewers` method, which fetches
  all preferred reviewers for all files in a change set, rather than
  limiting the number of reviewers per file to 1 as in
  `getPreferredReviewers`. This is necessary for `CheckInternal` to
  check that code reviewers with expertise over specific docs paths are
  counted as valid docs reviewers.
- Changes `getDocsReviewers`, `CheckInternal`, and
  `checkInternalDocsReviews` to take a slice of `PullRequestFile`s so
  they can apply the `getPreferredReviewers` logic.

* Respond to zmb3 feedback

- Fix spelling.
- Add log if admin reviewers are assigned.
- Deduplicate preferred reviewers.