Single Approval Paths
Add the ability for certain PRs to require a single approval.

A new setting has been added to the bot configuration file,
singleApproverPaths, that defines paths in a repository that
only require a single approver.

If all the files in a PR only match paths defined in
singleApproverPaths, then the PR only requires a single approval.
Otherwise, the PR requires the default number of approvals, which is
currently defined as 2.