libs/git: Handle "Z" timezone in dates
Use the standard `time.RFC3339` time format with `time.Parse()` so that
we support dates with a `Z` suffix instead of just a UTC offset. The
simple `Z` format is allowed by ISO8601 format so we should parse that.
It turns out that the `%cI` format passed to `git show` will return the
format with the `Z` suffix.

This fixes an error when running the changelog tool:

    parsing time "2024-09-25T02:35:11Z" as "2006-01-02T15:04:05-07:00": cannot parse "Z" as "-07:00"