libs/github: Make query date ranges consistent
Change the form of the github query date formatting to use `>=` instead
of `>` when just a start date is provided. The makes it consistent with
the range format (`a..b`) used when an end date is provided as the range
includes the first date. The changelog tool will need to be updated to
handle this by bumping the start dates by 1 second, but this should fix
an issue where an Enterprise commit for a PR that gets a release tag
ends up listed in two releases.