Add a workflow to check docs URLs
Like the workflow to check for necessary redirects to match renamed or
deleted docs pages, this workflow takes the path to a
`gravitational/teleport` clone. It finds mentions of Teleport docs site
paths in each file changed by a PR and ensures that each path has a
corresponding docs page or redirect source.