dependency review: disable comment summary
The job will fail if the comment it tries to post to the PR is
too large. This is fixed in newer versions of the action, but there
are other breaking changes to navigate if we want to update so the
quick fix for now is to disable comment summaries.