Add codeowners for /web dep updates
With the [dependabot.yaml
deprecation](https://github.blog/changelog/2025-04-29-dependabot-reviewers-configuration-option-being-replaced-by-code-owners/)
the suggested change is to add reviewers to the CODEOWNERS file
directly. Unfortunately this doesn't seem to give us the granularity to
select "only dependabot PRs" and so I think the next best thing is just
assign us to any PR that updates the package.json files (which is
usually just dependabot).

very open to alternative suggestions, and also we can maybe keep this
open and add the other configs for the go side of things as well