changelog: Support multiple changelog entries per PR
Allow a PR to have multiple changelogs, each following the standard
format of `"[Cc]hangelog: (<changelog>)"`. This was previously supported
in the shell script version of this so some people have been using it.

Update tests with a multi-changelog entry.

Fix the version in `go.mod` to remove the patch release and remove the
toolchain version. This way it only specifies the version of Go needed
to build this module and does not impose a specific toolchain version.