ci: Update golangci-lint to 2.3.0, run CI on workflow change
Update golangci-lint from 1.61.0 to 2.3.0 as 1.61.0 does not seem to
properly handle the later (1.23+) versions of Go.

Have the reusable-ci workflow run if that workflow file itself changes,
not just stuff in the tools directory. This makes sure that all tools
are kept up-to-date with the tooling.

Fix newly-surfaced lint errors in amplify-preview and env-loader. Ignore
errors in approvals-service as they are being fixed in another PR.