tools/common: Rework output for common test and lint (#386)
* tools/common: Remove CLI flag for colored lint output

Remove the `--out-format colored-line-number` as that is no longer
present on golangci-lint 2.0+, and if that later version is installed,
`make lint` fails.

* tools/common: Only use GitHub test output format if on GitHub

Only pass `--format github-actions` to `gotestsum` if we are running on
GitHub Actions. The output is rather verbose / noisy for locally running
`make test`.

* ci: Update golangci-lint to 2.3.0, run CI on workflow change

Update golangci-lint from 1.61.0 to 2.3.0 as 1.61.0 does not seem to
properly handle the later (1.23+) versions of Go.

Have the reusable-ci workflow run if that workflow file itself changes,
not just stuff in the tools directory. This makes sure that all tools
are kept up-to-date with the tooling.

Fix newly-surfaced lint errors in amplify-preview and env-loader. Ignore
errors in approvals-service as they are being fixed in another PR.