docpaths: Fix bug in identifying renames
The docpaths workflow mis-identifies docs path changes as requiring
redirects if those changes rename a regular page to a category index
page or vice versa. #380 addressed a similar bug, but assumed that the
GitHub API would classify changes that include this kind of renaming as
additions and deletions.

However, if the GitHub API classifies this kind of change as a rename,
the workflow incorrectly looks up a redirect source, even though the URL
path for the category index page is identical to the URL for the regular
page.

This change addresses the bug by editing the docpaths workflow to,
before checking for a redirect source of a renamed file, ensuring that
the previous and current URLs are not actually identical.