Simplify docpaths logic and fix a bug
Currently, the docpaths workflow addresses a corner case in which GitHub
marks a file rename as a deletion and an addition. However, there are
other possible ways GitHub may classify file renames. For example, if we
rename Page A to Page B, and incorporate content from Page C into Page
B, the GitHub API may mark Page A as a deletion and Page B as a rename
from Page C (rather than an addition).

As a result, in more complex renaming scenarios (e.g.,
gravitational/teleport#59329), the docpaths workflow can require that
the author add a redirect from a file that already exists, i.e., an
invalid redirect from one path to the same path.

This change accommodates these more complex rename scenarios. The
docpaths workflow checks if the path to check (the previous filename for
a renamed file or the filename for a deleted file) corresponds to a file
introduced by the PR, without assuming that this would be due to a
specific scenario.

Also use an enum for GitHub file statuses to prevent typos in string
values from introducing bugs.